Remove utils.chdir. NFC
This utility was only used in test code so move it there.
